Fix: 팔로우 언팔로우 버튼에 있던 함수 제거, UserInfo에 추가
