+++ src/components/UserInfo.jsx
@@ -1,13 +1,65 @@
 import React from 'react';
 import { Link, useParams } from 'react-router-dom';
+import axios from 'axios';
 import styles from './UserInfo.module.css';
 import FollowButton from './button/FollowButton';
 import UnFollowButton from './button/UnFollowButton';
+import { useState } from 'react';
+import { useEffect } from 'react';
 
 function UserInfo({ profileData }) {
     const getAccountName = localStorage.getItem('accountname');
     const { accountname } = useParams();
     const { myprofile } = useParams();
+    const getToken = localStorage.getItem('token');
+    const [state, setState] = useState(profileData.isfollow);
+    useEffect(() => {
+        setState(profileData.isfollow);
+    }, [profileData]);
+
+    const handleButton = () => {
+        if (profileData.isfollow) {
+            axios
+                .delete(
+                    `https://mandarin.api.weniv.co.kr/profile/${accountname}/unfollow`,
+                    {
+                        headers: {
+                            'Authorization': `Bearer ${getToken}`,
+                            'Content-type': 'application/json',
+                        },
+                    }
+                )
+                .then((res) => {
+                    console.log(res);
+                    const isFollow = res.data.profile.isfollow;
+                    setState(isFollow);
+                })
+                .catch((error) => {
+                    console.log(error);
+                });
+        } else {
+            axios
+                .post(
+                    `https://mandarin.api.weniv.co.kr/profile/${accountname}/follow`,
+                    {},
+                    {
+                        headers: {
+                            'Authorization': `Bearer ${getToken}`,
+                            'Content-type': 'application/json',
+                        },
+                    }
+                )
+                .then((res) => {
+                    console.log(res.data.profile.isfollow);
+                    const isFollow = res.data.profile.isfollow;
+                    setState(isFollow);
+                })
+                .catch((error) => {
+                    console.log(error);
+                });
+        }
+        console.log(state);
+    };
     return (
         <div className={styles.all_warpper}>
             <div className={styles.top_warpper}>
@@ -52,7 +104,10 @@ function UserInfo({ profileData }) {
                                 type="button"
                             ></button>
                         </li>
-                        <li className={styles.follow_button_warp}>
+                        <li
+                            className={styles.follow_button_warp}
+                            onClick={handleButton}
+                        >
                             {profileData.isfollow === false ? (
                                 <FollowButton />
                             ) : (
